Fix the more complex domain test.
The problem is that *serializing* directly the domain in the AST visitor's
stack creates the problem that we loose the connections between the and/or
nodes and its children.

Actually the connection is still there (we can retrieve it from the relation
of operators/count-of-arguments), but it's a bit more complex.

So the idea is to walk the AST and create a simpler tree structure and can
represent Odoo domains but simpler than full Python expressions, and that is
easier to create the domain.