Add the basis for the type-checker algorithm.
The type-checker algorithm is fully described in chapter 9 (but you should
also read chapter 8) of 'The Implementation of Functional Programming
Languages' by Simon L. Peyton Jones and others.

The core of the type-checker algorithm is done with an unification algorithm
in the terms of find a *substitution* phi that makes two types t1 and t2 the
same:

  to solve                   t1 = t2

  we find phi such that:     subtype phi t1 = subtype phi t2

In our first attempts we'll probably implement the algorithm as closely as
possible to the one in the Book.  But, it should be clear that we should cast
this type-checking problem in the language of the general term-unification
algorithm this branch is all about.