Allow to parse more complex type-expressions.
Now allows to parse, for instance, '(a -> b) -> c'.  We even allow
'f a -> g b' -- this required a trick to recast the previously
detect TypeVar('f') to a TypeCons; which in this case is actually
an application of higher-kind 'f'.

Now we can parse the S combinator.

Based on `ply`.

I included the output of the ply state machine (parser.out).