Extend the grammar to allow '(<op>)' to be a full expression.
Please ignore the changes in auto-generated files:

	modified:   xopgi/ql/lang/expressions/_exprtab.py
	modified:   xopgi/ql/lang/expressions/parser.out

We had created an artificial (and unneeded) mapping from operators to
functions names.  Let's be more like Haskell and let each operator to be its
own symbol.

So we don't need to 'apply' user-defined operators.  Everything is
user-defined, although some functions will be provided for the user.

In the grammar we gained the productions:

   expr : LPAREN DOT_OPERATOR RPAREN
        | LPAREN operator RPAREN

Which specifically allow an operator enclosed in parenthesis to be an
expression.

This allows (and solves for '(.) a b') the application of such operators (even
partial application) in more general way.