Ensure the tick infix application is left associative.
We have to make the token a little bit bigger, but since the ticks can only
enclose an identifier we just push to the lexer the what was before in the
parser.

We've made the infix tick to have less priority than composition and other
operators.  However there's bug there:

   'a ++ b `f` c'  should be eq. to '(a ++ b) `f` c' which is 'f((a ++ b) c)'

but the parser says is the same as:

   '(++) a (f b c)'