Test (and fix) number annotations.
Annotations are a meant to be used for concrete numbers.  But the parser
allows only to annotate a number with a string, char or identifier.

Semantically, we might be even make identifiers similar to a non-quoted string
so that it's easier to say 1@km than 1@"km".  The namespace for such
strings might defined elsewhere.