Resolve PADDING conflicts at the 'expr' level.
Having the rules:

   expr : PADDING expr
        | expr PADDING

Allowed us to parse entire strings (representing expressions) with padding at
the end or beginning:

   >>> parse('''
       a + b
   ... ''')
   Application(Application(Identifier('+'), Identifier('a')), Identifier('b'))

But I think has complicate the states of the parser.

I added a temporary standalone expression production ``st_expr`` and remove
the PADDING as part of the expressions.

Note that when we introduce `let` expressions will need the PADDING to cut
several expressions:

     let id x = x
         add x y = x + y
     in ...

This is a single expression but the PADDING breaks each *equation* in the
let.