Remove the 'lambda' keyword from the parser.
As said before, having two different ways to represent the same concept is a
bad idea.

It was even worse because I allowed the arrow to be used in keyword lambdas,
let's see it at its grandest splendor:

   \a -> lambda b -> b

which thanks to the Gods of PLY parses correctly; but... seriously?!

Therefore, the 'lambda' keyword is out.