Add a full description of the intended expression language.
Several doctest fail because of parsing errors.  Some are just issues with
wrong precedence or ambiguity of meaning.

A funny issue is the one where '+ 1' is being regarded a number instead of an
operator being applied:

    >>> parse('f a + 1')
    Application(Application(Identifier('+'), Application(Identifier('f'), Identifier('a'))), Literal(1, TypeCons('Number', [])))

    >>> parse('(f a) + 1')
    Application(Application(Identifier('f'), Identifier('a')), Literal(1, TypeCons('Number', [])))

Look at the tokens:

    >>> tokenize('f a + 1')
    [LexToken(IDENTIFIER,'f',4,0),
     LexToken(SPACE,' ',4,1),
     LexToken(IDENTIFIER,'a',4,2),
     LexToken(PLUS,'+',4,4),
     LexToken(BASE10_INTEGER,'1',4,6)]

    >>> tokenize('(f a) + 1')
    [LexToken(LPAREN,'(',4,0),
     LexToken(IDENTIFIER,'f',4,1),
     LexToken(SPACE,' ',4,2),
     LexToken(IDENTIFIER,'a',4,3),
     LexToken(RPAREN,')',4,4),
     LexToken(PLUS,'+',4,6),
     LexToken(BASE10_INTEGER,'1',4,8)]

In the second expression there's no SPACE between the ')' and '+'.  So the '+'
is the unary plus instead of the binary '... + ...'.  This is issue is
exacerbated when the '+' cannot be a unary operator:

   >>> parse('(f a) + c')
   Traceback ...