Type-check all AST constructs but Letrec.
Following the book at this point is very tricky.  Even though I'm doing the
most faithful translation from the Miranda code to Python, there are certain
things that simply don't belong here.

I use exceptions to communicate failure, so the 'reply' type is totally
removed.  This also implies that definitions with a FAILURE clause are not
explicitly stated.  But then the order of evaluation becomes important.

To illustrate:  The function `tcletrec` is the one used to type-check a
letrec, and it's defined like this:

    tcletrec :: type_env -> name_supply -> [vname] -> [vexp] -> vexp -> reply (subst, type_exp)
    tcletrec gamma ns xs es e
           = tcletrec1 gamma nsO nbvs e (tcl (nbvs ++ gamma) ns1 es)
                  where (nsO,ns') = split ns
                        (ns1 ,ns2) = split ns'
                        nbvs       = new_bvars xs ns2

Here `(tcl (nbvs ++ gamma) ns1 es)` may return a FAILURE.  So, in Python this
argument must be evaluated before doing the steps involved in the definition
of `tcletrec1`.

Notice that even in Miranda the `(tcl ...)` gets evaluated as if it were
strict-evaluation semantics because, in order to do the pattern-matching,
Miranda needs to execute the tcl to get either a FAILURE or an OK.