Optimize identities in sub-types.
The proof that 'subtype sidentity t = t' is easy:

If `t` is a type variable, 'subtype id t' is 'id(t.name)' which is 't'.

If `t` is a type cons on 0 subtypes, 'subtype id C' is C.  If `t` is a type
cons on several arguments, subtype is defined

    subtype id C a1 ... an = C (subtype id a1) ... (subtype id an)

The only thing we need to ensure is the each of the 'a's is a finite type.
Which we will left as an exercise ;).