Replace the implementation of a TypeEnvironment with a Mapping.
In the Book, the association list is a cheap choice and posses no risks nor
performance penalties (like copying the tail of a new list, because you can't
really update the tail of a list -- the cons can be shared.)

Previously we were just creating new lists by concatenating two lists
together.  This removes the risks that comes with sharing, but do have a toll
in memory.

Now we use ChainMap which do share things, but we were careful to introduce
all type hints to spot unintentional updates to environments (which are deemed
immutable mappings).