Add failing tests for all known parsing bugs.
I don't believe we should proceed solving each bug at a time; we may end with
a pile of patches instead of a robust parser.  These set of bugs should be
solved by taking a new approach to the parser.

Introducing a variation of a classical Term, Factor, Expression could work.

We don't have parser-related definitions like 'infixl $$ 9' in Haskell.  So
the programs don't instruct the parser.

The restriction about single names in 'let' is being kept in the doctest, but
lifted in the unit-tests so that which ever solution is easier (remove the
restriction or enforce it) we could always have a test to fix afterwards.