Merge the entire language.
Having the type language, expression language and program language separated
created some tension with the tokenizer needs of each.

For instance, the work to emit a SPACE and the newline rules in the type
languages needed for the its rule to work were not aligned with the same
tokens of the expression language.

Thus, trying to use the type language's rules embedded within the expression
language's rules to express whole programs was not straightforward.

At the very least we need a single lexer.

Decision:

Create a single coordinated 'parsers' module with the grammar for the full
language, but allow to create a *sub-parser* for the type languages and the
expression language.