Allow data constructors to take types as arguments.
There was an issues with a SPACE was being issued before the PIPE ('|')
operator; which makes the LARL parser to go ahead looking for another
argument.  So we must update t_SPACE to watch for '|'.

This issue was introduced in commit 15a6dcc01eff4f9e96e77b4a3ab0142570c7c6a1
when I added the PIPE token but forgot to handle the spacing issues.

Before this commit, 'data Tree a = Leaf a | Branch (Tree a) (Tree a)' would
also fail to recognize the arguments of Branch -- '(Tree a)'.  This is because
'_cons_args' was intentionally allowed to be a simple list of names.

This commit updates the nomenclature:

- 'args' or 'arguments' SHOULD NEVER be used in the LHS definitions.  'args'
  are always a thing of RHS.

- 'params' or 'parameters', or 'patterns' (for pattern-matching parameters)
  are to be used in LHS.

In the case of data type definitions there some point for debate since
'Branch (Tree a) (Tree a)' is NOT passing a argument to the Branch
constructor, but defining the constructor and stating the type of
its *parameters*.  But, at least for now, I keep the 'args' to the RHS.

With this in mind, '_cons_params' matches the list of type variables the type
constructor takes.  We allow only type variables (LOWER_IDENTIFIER).

'_cons_args' matches the list of arguments in the RHS of data types to match
the list arguments of each data constructor.  Any type expression can match.
Semantic analysis can disallow some cases.