Split type declarations from value definitions.
When parsing a program like::

   id :: a -> a
   id x = x

The type declaration can be *disconnected* from the definition itself.  This
is valid Haskell:

  id :: a -> a
  const :: a -> b -> a

  id x = x

  const a x = a

So the type declarations are not part of the definition.  It will be
type-checked later.  So we can simply return a TypeEnvironment with the type
annotation and later when we perform the type-checking it should be easier to
just augment the builtins with the new environment.

We extracted TypeScheme and TypeEnvironment from the typecheck module.  Now
they belong to the type language.