Solve parsing of expressions with COLON and user operators.
It seems that to solve the problem of having right and left assoc operators in
the same level of precedence, I would have to allow for::

    expr_termX : expr_temrX infix_operatorX expr_termX
               | expr_term_X+1

    infix_operatorX : infixr_operatorX | infixl_operatorX

But also I would have to know which rule matched when reducing in with the
`p_`function.  Also this function would be responsible of doing the *right
thing* with associativity by looking at a table that would map operators to
associativity.

For now, I put COLON alone in the level 1, which makes it to take less
precedence than any other operator but infix `functions`.