Add tests of unparsable programs I'd like to be parsed.
Because of the builtins.fl I'm making.  The 'builtins.fl' would our gold
test.

It's kind of the Preamble, but with missing definitions for values I'm going
to implement in Python directly.

But some things like the Maybe and Either data types can be implemented
directly.

I would like to have comments.  Maybe a with a preprocessor.