Extract the representation of a ConcreteLet expression.
It seems that we might to need to hold converting the patterns of a let
expression to a lambda expression in order to be able to do pattern matching.

Now ConcreteLet is the way to build the final Let expression.

For the time being, we keep producing the Let/Letrec expression in the parser
once the 'let' expression is finished parsing.

Pattern Matching may require to enlarge such construction.  As far as I can
gather from the book 'The implementation of functional programming languages';
pattern matching require some knowledge of the types (sum or product types).

So far, the only type-related feature of the parser was the builtin knowledge
about the type of literals.