Implement basic transformations for pattern matching.
While trying to implement the pattern matching, some things come to my
attention.

I was trying to compile each equation separately, but I then came to an
invalid translations.  Example:

    count []   = 0
    count x:xs = 1 + (count xs)

I was getting something like:

   count = \.lst -> ((\.nil -> 0) (<Match Nil> .lst))
                    `:OR:`
                    ((\.xs -> 1 + (count .xs)) (<Extract 2 from :> .lst))
                    `:OR:`
                    :NO_MATCH_ERROR:

This is semantically incorrect because the argument '.nil' is never evaluated
in the lambda's body, and, consequently, the application '<Match Nil> .lst' is
not evaluated.  Which means that this version of 'count' always returns 0.

So I need to take a step back:

   count = \.lst -> ((\[] -> 0) .lst) `:OR:`
                    ((\(x:xs) -> 1 + (count xs)) .lst) `:OR:`
                    :NO_MATCH_ERROR:

But those inner lambdas are pattern-matching lambdas; and they must evaluate
'.lst' as much as they need to perform the pattern matching.

Fixing the issues took some refactoring.  Lazy pattern evaluation, but at the
same time strict order of evaluation of several equations, require to take the
set of equations as whole.

The 'correct' translation would be:

    count = \.lst -> (<Match Nil> .lst 0)
                     `:OR:`
                     (<Extract 1 from :> .lst (\x -> (<Extract 2 from :> .lst (\xs -> 1 + (count xs)))))
                     `:OR`:
                     :NO_MATCH_ERROR:

The types of the match/extract function change to take the values (lambdas in
Extract; and AST in Match).  The semantics are to evaluate the argument (.lst)
to perform the pattern matching.

This commit is the beginning of such a refactor.  But, now simpler functions
like 'id x = x' get translated to:

   id = \.id_arg0 -> ((\x -> x) .id_arg0) `:OR:` :NO_MATCH_ERROR:

The broken tests (test_basic_letexpr, test_local_definitions) test the
resulting AST.  I'm leaving them broken because I intend to fix the
translation.  I also add a new (failing) test for nested patterns.