Reorganize the package to differentiate the AST modules.
There's still some refactoring to do.  Example: The FunctionDefinition is a
transformation from general AST to a restricted form we'll call Enriched
Lambda Calculus.  Also Match, Select and Extract are *values* (or identifiers)
outside the scope of the AST.

It's not sufficient to have a LCNode, because, for instance, a Lambda can
contain ConcreteLet; only after transformation all nodes are guaranteed to be
LC nodes.

To make most tests pass, we modify the typechecker to deal with ConcreteLet
expressions.