Correct the return type of FunctionDefinition.compile.
In fact, since we have a single tree of AST nodes, we cannot identify *by
type* if a node is only composed of nodes from the lambda calculus.

For instance the expression '\lst -> let head x:xs = x in head lst'; is not in
lambda calculus even if its top-level node is a Lambda abstraction.