Split the AST from Intermedite Language.
The language AST allows for constructions that are not part of the
intermediate language.

The class AST marks nodes of user-facing language.  AST nodes can contain
other AST nodes.

The class ILC marks nodes that are subject to compilation/interpretation.  ILC
can contain other ILC nodes.

As noted by the Dual class, there are nodes that part of both languages.
However, there's a translation API to get a ILC node from an AST.  In the dual
nodes, this is the identity function.