Fix incorrect type annotation for 'then'.
This was introduced in af6634e7a689461710bcdff7f9b63707dfac6358.

Detecting this issue was harder than needed because our language has poor
support for error reporting.  Also we allow for duplicated type annotations.

To reduce the change of such errors later on, I also hardened the collection
of builtins.  This also uncovered that product types Select must not be used
for type-checking.