Avoid repeating names in patterns.
Repeating the same would make our translation a bit harder.  Something like
'f x x = ...'; would have to be translated to

   f .a1 .a2
     | .a1 == .a2 = ...

And we don't have guards.  It will also imply that the values must implement
Eq; so it places an implicit constraint of the type of the arguments.

We exclude the '_' name which is not considered bound.  That's why we also had
to change the typechecker to consider this identifier a fresh identifier in
the expression.