Add test that should fail when we move beyond HM type inference.
This test shows that our type-inference algorithm is following HM system;
which make it put a single non-generalized unification var to the argument
'get'; that's why it fails to infer a type for 'get' -- it would not be able
to unify Num with Char.

When we type it ourselves -- that's one of the minor extensions we've
implemented to HM; it checks the type correctly.

This test will fail when we move beyond HM, because the type checker should be
able to deal with 'f get = ...' without the local type annotation.

The road to type-systems was (according to
http://lhc-compiler.blogspot.com/2016/09/haskell-suite-type-inference.html):
first HM [Damas1982], then bidirectional type inference [Pierce2000], followed
by boxy types [Vytiniotis2006] and finalizing with OutsideIn(X) [OutsideInX]

I don't know yet which system is best to our needs; but I'm beginning to
believe it should support extensible type classes and/or records.