Add a *concrete* implementation of the type expression language.
This is the first attempt to use Lark as the parsing framework.  Since the
Lark grammar allows for repetition of non-terminals and terminals by using +
and *, some are more compact and easier to understand.

Compare, for instance, the rule of type_scheme:

  type_scheme : KEYWORD_FORALL identifier identifier* "."

to the previous:

  type_scheme : empty
  type_scheme : KEYWORD_FORALL SPACE _type_scheme_generics
  _type_scheme_generics : _identifier SPACE _type_scheme_generics
  _type_scheme_generics : _identifier DOT_OPERATOR

There's still however limitations -- afterall is LALR(1).  The current
`type_expr_parse` fails to parse unconstrained 'Maybe a' or 'Maybe a -> a'.

But with Lark we could switch to Early parsing:

   >>> lark = Lark.open("xotl/fl/parsers/grammar.lark",
   ...                  lexer="standard", propagate_positions=True,
   ...                  start="type_expr", debug=True)
   >>> print(lark.parse('Maybe a').pretty())
   type_expr
     type_term
       type_app_expression
         type_factor
           type_cons	Maybe
         type_factor
           type_variable	a

   >>> print(lark.parse('Eq a => Maybe a').pretty())
   type_expr
     type_constraints
       type_constraint
         type_cons	Eq
         type_variable	a
       =>
     type_term
       type_app_expression
         type_factor
           type_cons	Maybe
         type_factor
           type_variable	a

However, I haven't decided to switch to Early... There's a simple solution by
allowing non-legal sentences and then raising the error while trying to build
the AST.

Another positive note is that previously failing type-expressions are now
parsed correctly:

   >>> type_expr_parser.parse("forall a. Eq a, Ord a => Maybe a")
   Tree(type_expr, ...)

== The cons ==

The documentation of Lark is not that good.  You have to go to the examples
and source code.

== Changes in the treatment of spaces ==

I'm ignoring inline spaces.  I'm planning the application is just 'expr
expr'.

One of the things I liked about Lark is that is more readable than PLY.
There's an indenter module which

   Provides Indentation services for languages with indentation similar to
   Python

So, maybe I could make use of it, or copy some of it to deal with our
indentation issues.

== Test failures ==

Almost all tests are bound to fail since I'm not building the AST.  But is
good to know which things I'm forgetting but looking at failing tests.