Use an Early parser.
Early may be slower, but for the purposes of this module, where compiling
large programs is not expected, restricting ourselves to LALR may simply make
the development slower and for no gain.

Since I want to move as quickly as possible to a complete language
specification with type check/inference, the parser is becoming less
important.  I hope Early will prove good enough.