Advance the expression language using lark.
I have just written the simplest of the expressions, not let expression, or
data type, etc; just "arithmetic" expressions.

Application has become a *list-like* element of tree because, otherwise I
would get the wrong association, i.e: the tree for "a b c", would be
equivalent to "a (b c)", and we expect "(a b) c".

I'm using heavily the Lark feature of hiding concrete elements of the tree (by
using "_" in rules).  Without it parsing "a f g" gets a tree like:

    >>> print(lark.parse(r"a f g").pretty())
    expr
      expr_term0
        expr_term1
          expr_term2
            expr_term6
              expr_term7
                expr_term9
                  expr_factor
                    application
                      expr_identifier	a
                      expr_identifier	f
                      expr_identifier	g

And with it:

    >>> print(expr_lark.parse(r"a f g").pretty())
    expr
      application
        expr_identifier	a
        expr_identifier	f
        expr_identifier	g

All expr_termN levels are hidden unless they actually add some value (the
operator):

    >>> print(expr_lark.parse(r"a f >>= g").pretty())
    expr
      expr_term2
        application
          expr_identifier	a
          expr_identifier	f
        infixl_operator_2	>>=
        expr_identifier	g