Mark next version as 0.3.0.
Skip release 0.2.0.  In fact, these are not releases per-se; for there's
nothing complete yet.  Let's make a plan:

- 0.3.0  Parser complete.
- 0.4.0  Type checker/inference complete.
- 0.5.0  Runtime complete.

All those releases should have all the planned features including:

- Pattern matching
- Type classes and instances
- ADTs

There are some features which I don't know if we need:

- Record types and sub-typing.

Those could be included in any of the releases, but the release could be done
without them.