Avoid generating COMMENT in the hypothesis generated samples.
Since the regular expression of COMMENT allows it at the beginning and end of
the source-code, hypothesis is generating comments at any point:

    'a  -- alskdj'

But comments in generated source code adds no value to test our parser.

In fact, several things in the AST (like checking TypeVariables and usage of
constraints in the constrained types) make the parse of generated randomly
data of little usage.  We need better ways to generate this kind of data.